+++ lib/models/expense.dart
@@ -2,12 +2,12 @@ class Expense {
   final String name;
   final double amount;
   final DateTime createdAt;
-  final String expenseType; // 'Need' or 'Desire'
+  final String? expenseType; // 'Need' or 'Desire'
 
   Expense({
     required this.name,
     required this.amount,
     required this.createdAt,
-    required this.expenseType,
+    this.expenseType,
   });
 }