+++ lib/screens/home_page.dart
@@ -17,28 +17,12 @@ class _HomePageState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: Text(widget.title),
-      ),
+      appBar: AppBar(backgroundColor: Theme.of(context).colorScheme.inversePrimary, title: Text(widget.title)),
       body: Padding(
         padding: const EdgeInsets.all(16.0),
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            if (expenses.isEmpty)
-              const Center(
-                child: Padding(
-                  padding: EdgeInsets.all(32.0),
-                  child: Text(
-                    'No expenses yet. Add your first expense!',
-                    style: TextStyle(fontSize: 16.0, fontStyle: FontStyle.italic),
-                  ),
-                ),
-              )
-            else
-              ExpenseTable(expenses: expenses),
-          ],
+          children: <Widget>[if (expenses.isEmpty) _buildEmptyState() else ExpenseTable(expenses: expenses)],
         ),
       ),
       floatingActionButton: FloatingActionButton(
@@ -58,6 +42,36 @@ class _HomePageState extends State<HomePage> {
         },
         tooltip: 'Add Expense',
         child: const Icon(Icons.add),
+      ),
+    );
+  }
+
+  Center _buildEmptyState() {
+    return Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Text(
+            'No expenses yet. Add your first expense!',
+            style: TextStyle(fontSize: 16.0, fontStyle: FontStyle.italic),
+          ),
+          const SizedBox(height: 16.0),
+          ElevatedButton(
+            child: Text('Add Expense'),
+            onPressed: () {
+              showDialog(
+                context: context,
+                builder: (BuildContext context) {
+                  return ExpenseInputTypeDialog(
+                    onExpenseAdded: () {
+                      setState(() {});
+                    },
+                  );
+                },
+              );
+            },
+          ),
+        ],
       ),
     );
   }